# src/api/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Medicamenta.me Public API",
    "version": "1.0.0",
    "description": "RESTful API for healthcare partner integrations. Manage patients, medications, adherence tracking, and more.",
    "contact": {
      "name": "API Support",
      "email": "api-support@medicamenta.me",
      "url": "https://docs.medicamenta.me"
    },
    "license": {
      "name": "Proprietary",
      "url": "https://medicamenta.me/terms"
    }
  },
  "servers": [
    {
      "url": "https://us-central1-{project-id}.cloudfunctions.net/api",
      "description": "Production"
    },
    {
      "url": "http://localhost:5001/{project-id}/us-central1/api",
      "description": "Local Development"
    }
  ],
  "security": [
    {
      "ApiKeyAuth": []
    },
    {
      "BearerAuth": []
    }
  ],
...
  "paths": {
...
    "/v1/patients": {
      "post": {
        "tags": ["Patients"],
        "summary": "Create patient",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatientInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Patient created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Patient"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "get": {
        "tags": ["Patients"],
        "summary": "List patients",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20,
              "maximum": 100
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 0
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["active", "inactive", "deleted"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of patients",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Patient"
                      }
                    },
                    "pagination": {
                      "$ref": "#/components/schemas/Pagination"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/patients/{id}": {
      "get": {
        "tags": ["Patients"],
        "summary": "Get patient",
        "parameters": [
          {
            "$ref": "#/components/parameters/PatientId"
          }
        ],
        "responses": {
          "200": {
            "description": "Patient details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Patient"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "patch": {
        "tags": ["Patients"],
        "summary": "Update patient",
        "parameters": [
          {
            "$ref": "#/components/parameters/PatientId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatientInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Patient updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Patient"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Patients"],
        "summary": "Delete patient",
        "parameters": [
          {
            "$ref": "#/components/parameters/PatientId"
          }
        ],
        "responses": {
          "204": {
            "description": "Patient deleted"
          }
        }
      }
    },
    "/v1/medications": {
      "post": {
        "tags": ["Medications"],
        "summary": "Create medication",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MedicationInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Medication created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Medication"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["Medications"],
        "summary": "List medications",
        "parameters": [
          {
            "name": "patientId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["active", "completed", "deleted"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of medications",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Medication"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key for authentication"
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token from /v1/auth/token endpoint"
      }
    },
    "parameters": {
      "PatientId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "Patient ID"
      }
    },
    "schemas": {
...
      "PatientInput": {
        "type": "object",
        "required": ["name", "dateOfBirth"],
        "properties": {
          "name": {
            "type": "string",
            "example": "João Silva"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "phone": {
            "type": "string"
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date"
          },
          "gender": {
            "type": "string",
            "enum": ["M", "F", "Other"]
          },
          "medicalConditions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "allergies": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "Patient": {
        "allOf": [
          {
            "$ref": "#/components/schemas/PatientInput"
          },
          {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "status": {
                "type": "string",
                "enum": ["active", "inactive", "deleted"]
              },
              "createdAt": {
                "type": "string",
                "format": "date-time"
              },
              "updatedAt": {
                "type": "string",
                "format": "date-time"
              }
            }
          }
        ]
      },
      "MedicationInput": {
        "type": "object",
        "required": ["patientId", "name", "dosage", "frequency"],
        "properties": {
          "patientId": {
            "type": "string"
          },
          "name": {
            "type": "string",
            "example": "Losartana"
          },
          "dosage": {
            "type": "string",
            "example": "50mg"
          },
          "frequency": {
            "type": "string",
            "enum": ["daily", "twice_daily", "three_times_daily", "custom"]
          },
          "times": {
            "type": "array",
            "items": {
              "type": "string",
              "pattern": "^([01]?[0-9]|2[0-3]):[0-5][0-9]$"
            },
            "example": ["08:00", "20:00"]
          },
          "instructions": {
            "type": "string"
          }
        }
      },
      "Medication": {
        "allOf": [
          {
            "$ref": "#/components/schemas/MedicationInput"
          },
          {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "adherenceRate": {
                "type": "number",
                "format": "float"
              },
              "status": {
                "type": "string"
              }
            }
          }
        ]
      },
...
      "Pagination": {
        "type": "object",
        "properties": {
          "limit": {
            "type": "integer"
          },
          "offset": {
            "type": "integer"
          },
          "total": {
            "type": "integer"
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "details": {
                "type": "object"
              },
              "timestamp": {
                "type": "string",
                "format": "date-time"
              }
            }
          }
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Bad request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "NotFound": {
        "description": "Not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    }
  }
}